Add support for checkboxes fields and options to acf_field_fallback shortcode
This commit adds a new shortcode, [acf_checkboxes_list], that allows displaying the values of an ACF checkboxes field in a list format. It also adds the ability to specify optional prefix and suffix text for the [acf_field_fallback] shortcode, as well as the option to disable capitalization of the field value.

The README.md file has been updated to include documentation on these changes and how to use them.